+++ zarrs_metadata/CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Add `Configuration::to_typed()`
+- Add `Metadata{V3,V2}::to_typed_configuration()`
+- Add `EmptyConfiguration`
+
 ## [0.7.1] - 2026-01-13
 
 ### Added